% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_fns.R
\name{ests_std}
\alias{ests_std}
\title{Compute standardized Rogan-Gladen prevalence estimates

Corresponds to \eqn{\hat{\pi_{SRG}}} and \eqn{\hat{V_{\pi, SRG}}} in the manuscript}
\usage{
ests_std(
  sample,
  sigma_e_hat,
  sigma_p_hat,
  n_1,
  n_2,
  n_3,
  vars_std,
  variance = FALSE
)
}
\arguments{
\item{sample}{A dataframe with data from the sample, including the stratum_props (gamma_js) which will be used in standardization}

\item{sigma_e_hat}{Estimated sensitivity}

\item{sigma_p_hat}{Estimated specificity}

\item{n_1}{Sample size for sensitivity validation dataset}

\item{n_2}{Sample size for specificity validation dataset}

\item{n_3}{Sample size for main study from population}

\item{vars_std}{The names of variables to standardize over, corresponding to columns in sample}

\item{variance}{Should variance be estimated? (default = TRUE)}
}
\value{
Estimated prevalence; optionally, a vector also including variance estimate
}
\description{
Compute standardized Rogan-Gladen prevalence estimates

Corresponds to \eqn{\hat{\pi_{SRG}}} and \eqn{\hat{V_{\pi, SRG}}} in the manuscript
}
\examples{
ests_std(sample, .98, .99, 40, 273, 3000, c("age", "race"), variance = TRUE)
}
