% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_fns.R
\name{ests_std_model}
\alias{ests_std_model}
\title{Compute logistic regression standardized Rogan-Gladen estimates}
\usage{
ests_std_model(
  sample,
  stratum_props,
  sigma_e_hat,
  sigma_p_hat,
  n_1,
  n_2,
  n_3,
  vars_std,
  mod_formula,
  variance = FALSE
)
}
\arguments{
\item{sample}{A dataframe with data from the sample, including the stratum_props (\eqn{\gamma_j} values) which will be used in standardization}

\item{stratum_props}{the stratum_props dataframe. note that the vars_std need to be the first columns from the left}

\item{sigma_e_hat}{Estimated sensitivity}

\item{sigma_p_hat}{Estimated specificity}

\item{n_1}{Sample size for sensitivity validation dataset}

\item{n_2}{Sample size for specificity validation dataset}

\item{n_3}{Sample size for main study from population}

\item{vars_std}{The names of variables to standardize over, corresponding to columns in sample}

\item{mod_formula}{An object of class "formula" that symbolically describes the model to be fitted}

\item{variance}{Should variance be estimated? (default = TRUE)}
}
\value{
Estimated prevalence; optionally, a vector also including variance estimate
}
\description{

}
\examples{
ests_std(sample, stratum_props, .98, .99, 40, 273, 3000, vars_std = c("age", "race"), 
                   mod_formula = formula("x ~ age + race"), variance = TRUE)
}
